setup-ivi-fstab: use correct case
Apparently, mount is case-sensitive. The UUID for normal file-systems like ext4
must be lower case, just like it is in /dev/disk/by-uuid/. But the short FAT-FS
ID has to be all-capitals. Take this into account in 'setup-ivi-fstab'.

Also add a couple more checks to make sure the installer framework variables
that we use exist.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>