installerfw-sh-functions: do not omit the last line
When we are restoring the installer-framework environment in the
'installerfw_restore_env()', we read the '/etc/installerfw-environment' file
line-by line, and export variables.

However, if the last line does not contain the newline character, we ignore it.
So we miss the last variable and do not export it.

This patch fixes the problem, which also fixes TIVI-2652.

Change-Id: I47f560e32a6951b492b0e430a769c4dfc29e5cb5
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>